requirements.txt
pytest==8.3.3
pytest-asyncio==0.24.0
requests==2.32.3
pyyaml==6.0.2
llama-cpp-python>=0.2.0


